pyproject.toml
[tool.poetry]
name = "apt-management-backend"
version = "0.1.0"
description = ""
authors = ["msm9401 <anstmdals94@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "apt_management_backend"}]

[tool.poetry.dependencies]
python = "^3.8"
django = "4.1"
django-environ = "^0.9.0"
requests = "^2.28.1"
xmltodict = "^0.13.0"
djangorestframework = "^3.14.0"
django-rest-knox = "^4.2.0"
locust = "^2.14.2"
django-debug-toolbar = "^3.8.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
